streamlit/dockerfiles/requirements.txt
streamlit==1.26.0
requests==2.31.0
grpcio==1.57.0
grpcio-tools==1.57.0
protobuf==4.23.3